seperate main.cpp into two files
